% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_utils.R
\name{discard_burn_in.expansionsMCMC}
\alias{discard_burn_in.expansionsMCMC}
\title{Discard burn in iterations}
\usage{
\method{discard_burn_in}{expansionsMCMC}(x, proportion = NULL, k_it = NULL, ...)
}
\arguments{
\item{x}{expansionsMCMC object}

\item{proportion}{(Optional) The proportion of iterations to be discarded. Either 'proportion' or 'k_it' must be provided}

\item{k_it}{(Optional) The number of iterations to be discarded. Either 'proportion' or 'k_it' must be provided}
}
\value{
expansionsMCMC object with iterations designated as burn in discarded
}
\description{
Discard burn in iterations
}
