% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions_infer.R
\name{expansions_infer}
\alias{expansions_infer}
\title{Run rjmcmc inference on the provided preprocessed phylogeny, and using provided prior distributions
This is a barebones method for advanced usage. For standard usage refer to `run_expansion_inference`}
\usage{
expansions_infer(
  pre,
  prior_i,
  prior_N,
  prior_t_mid_given_N,
  prior_t_mid_given_N.sample,
  prior_K_given_N,
  prior_K_given_N.sample,
  prior_t_given_N,
  prior_t_given_N.sample,
  concentration,
  n_it = 1e+06,
  thinning = 1,
  init = NULL,
  debug = FALSE
)
}
\arguments{
\item{pre}{Preprocessed phylogeny}

\item{prior_i}{Prior likelihood on number of expansions}

\item{prior_N}{Prior likelihood on background population size}

\item{prior_t_mid_given_N}{Prior likelihood on expansion time to midpoint given background population size}

\item{prior_t_mid_given_N.sample}{Sampling function for prior on time to midpoint given background population size}

\item{prior_t_given_N}{Prior likelihood on expansion time given background population size}

\item{prior_t_given_N.sample}{Sampling function for prior on expansion timegiven background population size}

\item{concentration}{concentration parameter for the dirichlet prior on expansion membership probabilities}

\item{n_it}{number of MCMC iterations}

\item{thinning}{mcmc output thinning}

\item{init}{initial state for the mcmc chain}

\item{debug}{debug flag}

\item{prior_K_mid_given_N}{Prior likelihood on expansion carrying capacity given background population size}

\item{prior_K_mid_given_N.sample}{Sampling function for prior on expansion carrying capacity given background population size}
}
\value{
a list of MCMC states
}
\description{
Run rjmcmc inference on the provided preprocessed phylogeny, and using provided prior distributions
This is a barebones method for advanced usage. For standard usage refer to `run_expansion_inference`
}
