% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_utils.R
\name{expansionsMCMC}
\alias{expansionsMCMC}
\title{Constructs object of class expansionsMCMC}
\usage{
expansionsMCMC(
  phylo_preprocessed,
  priors,
  model_data,
  expansion_data,
  metadata
)
}
\arguments{
\item{phylo_preprocessed}{an object of type preprocessedPhy}

\item{priors}{an object of type priorList}

\item{model_data}{dataframe containing model data as returned by mcmc2data.frame}

\item{expansion_data}{dataframe containing expansion data as returned by mcmc2data.frame}

\item{metadata}{mcmc metadata}
}
\value{
an object of type expansionsMCMC
}
\description{
Constructs object of class expansionsMCMC
}
