% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_coalescent_loglh}
\alias{exponential_coalescent_loglh}
\title{Compute likelihood of a realisation of an exponential growth coalescent.}
\usage{
exponential_coalescent_loglh(
  sampling_times,
  coalescent_times,
  lambda,
  pop_size
)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{coalescent_times}{times of coalescent events.}

\item{lambda}{growth rate.}

\item{pop_size}{final effective pop_size at time of most recent sample.}
}
\value{
the log likelihood for the given parameters.
}
\description{
Compute likelihood of a realisation of an exponential growth coalescent.
}
