% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_coal.R
\name{extract_lineage_times}
\alias{extract_lineage_times}
\title{Extracts Lineage times from a parent phylogeny based on provided divergence MRCA nodes and times.}
\usage{
extract_lineage_times(phylo.preprocessed, div.MRCA.nodes, div.times)
}
\arguments{
\item{phylo.preprocessed}{preprocessed phylogeny}

\item{div.MRCA.nodes}{labels of MRCA nodes of diverging lineages}

\item{div.times}{absolute divergence times}
}
\value{
A list containing lists of sampling time vectors sam.times and coalescent time vectors coal.times for each lineage.
}
\description{
Extracts Lineage times from a parent phylogeny based on provided divergence MRCA nodes and times.
}
