% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenous_coal.R
\name{homogenous_coal.log_lh}
\alias{homogenous_coal.log_lh}
\title{Compute likelihood of a realisation of a homogeneous coalescent.}
\usage{
homogenous_coal.log_lh(sampling_times, coalescent_times, pop_size)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{pop_size}{Effective population size \code{Neg}.}
}
\value{
the log likelihood for the given parameters.
}
\description{
Compute likelihood of a realisation of a homogeneous coalescent.
}
