% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenous_coal.R
\name{homogenous_coal.simulate}
\alias{homogenous_coal.simulate}
\title{Simulate homogenous coalescent,}
\usage{
homogenous_coal.simulate(sampling_times, pop_size)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{pop_size}{Effective population size \code{Neg}.}
}
\value{
A list consisting of the simulate coalescent times \code{coalescent_times} and the log-likelihood of the simulation \code{log_likelihood}.
}
\description{
Simulate homogenous coalescent,
}
