% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomogenous_coal.R
\name{inhomogenous_coal.log_lh}
\alias{inhomogenous_coal.log_lh}
\title{Compute likelihood of a realisation of an inhomogeneous coalescent.}
\usage{
inhomogenous_coal.log_lh(
  sampling_times,
  coalescent_times,
  Neg.rate,
  Neg.rate.int
)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{coalescent_times}{times of coalescent events.}

\item{Neg.rate}{function equal to 1/Neg(t).}

\item{Neg.rate.int}{integral of 1/Neg(t) from \code{t} to \code{t+s}.}
}
\value{
the log likelihood for the given parameters.
}
\description{
Compute likelihood of a realisation of an inhomogeneous coalescent.
}
