% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomogenous_coal.R
\name{inhomogenous_coal.simulate}
\alias{inhomogenous_coal.simulate}
\title{Simulate inhomogenous coalescent,}
\usage{
inhomogenous_coal.simulate(
  sampling_times,
  Neg.rate,
  Neg.rate.int,
  Neg.rate.int_inv
)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{Neg.rate}{function equal to 1/Neg(t).}

\item{Neg.rate.int}{integral of 1/Neg(t) from \code{t} to \code{t+s}.}

\item{Neg.rate.int_inv}{function that inverts the integral of 1/Neg(t) from \code{t} to \code{t+s} for \code{s}.}
}
\value{
A list consisting of the simulate coalescent times \code{coalescent_times} and the log-likelihood of the simulation \code{log_likelihood}.
}
\description{
Simulate inhomogenous coalescent,
}
