% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{inv_rates}
\alias{inv_rates}
\title{Numerically invert rates}
\usage{
inv_rates(func, N, dfun = NULL)
}
\arguments{
\item{func}{function to invert}

\item{N}{function value}

\item{dfun}{analytic derivative for quasi newtonian scheme}
}
\value{
value of inverse function
}
\description{
Numerically invert rates
}
