% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_utils.R
\name{mcmc2data.frame}
\alias{mcmc2data.frame}
\title{Converts MCMC output into two data frames, one conaining global model parameters and one containing expansion data}
\usage{
mcmc2data.frame(o)
}
\arguments{
\item{o}{MCMC output}
}
\value{
A list with names mcmc.df and event.df. MCMC dataframe contains iteration numbers, 
        model indicator at given iteration, background population size and likelihood and prior values
        event.df contains rows describing different expansions. Columns consist of which iteration does an expansion belong to, 
        and the associated t_mid/K/time/branch/probability values.
}
\description{
Converts MCMC output into two data frames, one conaining global model parameters and one containing expansion data
}
