% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_utils.R
\name{plot.expansionsMCMC}
\alias{plot.expansionsMCMC}
\title{Plot results}
\usage{
\method{plot}{expansionsMCMC}(
  x,
  ...,
  mode = c("summary", "modes", "persistence", "traces", "mtraces", "popfunc"),
  k_modes = NULL,
  correlates = list(),
  corr_axis_title = list(),
  corr_legend_title = list(),
  gt.K = NULL,
  gt.t_mid = NULL,
  gt.time = NULL,
  branch = NULL,
  t_max = NULL,
  tree_scale = NULL
)
}
\arguments{
\item{x}{expansionsMCMC object}

\item{...}{further arguments passed to plot}

\item{mode}{one of "summary" - produces a summary plot for this chain, "modes" - plot marginals for top k posterior probability branches, "persistence" - plot posterior probability of expansion tip membership, "traces" - plot traces for N, number of expansions, and branches , "mtraces" - plot traces for expansion parameters for top k posterior probability branches, "popfunc" - plot posterior population functions.}

\item{k_modes}{number of highest posterior probability branches to generate plots for used by "modes", "mtraces", "popfunc"}

\item{correlates}{(Optional) correlates to plot alongside persistence matrix used by "persistence" plot}

\item{corr_axis_title}{(Optional) axis title for correlate map used by "persistence" plot}

\item{corr_legend_title}{(Optional) legend title for correlate map used by "persistence" plot}

\item{gt.K}{true K values if plotting for diagnostic purposes}

\item{gt.t_mid}{true t_mid values if plotting for diagnostic purposes}

\item{gt.time}{true expansion time values if plotting for diagnostic purposes}

\item{branch}{which branch to generate "popfunc" plot for}

\item{t_max}{upper x-axis limit for "popfunc" plot}

\item{tree_scale}{time unit axis label}
}
\description{
Plot results
}
