% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_utils.R
\name{plot_dim_panel}
\alias{plot_dim_panel}
\title{Plots a panel summarising the inferred number of expansions and base population size}
\usage{
plot_dim_panel(mcmc.df, prior_N = NULL)
}
\arguments{
\item{mcmc.df}{mcmc.df returned by mcmc2data.frame}

\item{prior_N}{(Optional) Background population size prior, mutually exclusive with passing list of priors. If either is supplied priors will be overlayed in plotting.}
}
\value{
a panel containing histograms and traces for the inferred number of expansions and the base population size
}
\description{
Plots a panel summarising the inferred number of expansions and base population size
}
