% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomogenous_coal.R
\name{plot_exp_growth}
\alias{plot_exp_growth}
\title{Simulate and plot a coalescent process for an exponentially growing population.}
\usage{
plot_exp_growth(sam, lambda, N)
}
\arguments{
\item{lambda}{growth rate.}

\item{N}{final population size at time 0.}

\item{sampling_times}{times of leaves.}
}
\value{
A list consisting of the simulate coalescent times \code{coalescent_times} and the log-likelihood of the simulation \code{log_likelihood}.
}
\description{
Simulate and plot a coalescent process for an exponentially growing population.
}
