% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{plot_structured_tree}
\alias{plot_structured_tree}
\title{Plots a structured tree with different lineages distinguished by different colours, and divergence events highlighted.}
\usage{
plot_structured_tree(tree, n_lineages)
}
\arguments{
\item{n_lineages}{number of lineages}

\item{sampling_times}{a tree including divergence events}
}
\value{
a list with the Newick string corresponding to the tree, and a list of leaf colouring assignments
}
\description{
Plots a structured tree with different lineages distinguished by different colours, and divergence events highlighted.
}
