% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{plot_tree_freq}
\alias{plot_tree_freq}
\title{Plots tree branch frequency histogram}
\usage{
plot_tree_freq(
  mcmc.df,
  event.df,
  pre,
  prior_t_given_N = NULL,
  highlight_node = NULL
)
}
\arguments{
\item{mcmc.df}{mcmc.df returned by mcmc2data.frame}

\item{event.df}{event.df returned by mcmc2data.frame}

\item{pre}{Preprocessed phylogeny}

\item{prior_t_given_N}{(Optional) Expansion time prior. If supplied prior will be overlayed.}

\item{highlight_node}{(Optional) Node to highlight}
}
\value{
a plot object
}
\description{
Plots tree branch frequency histogram
}
