% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_utils.R
\name{rescale_to_exponential}
\alias{rescale_to_exponential}
\title{Rescale a homogenous process to one with exponentially growing population.}
\usage{
rescale_to_exponential(c_t, beta, t0)
}
\arguments{
\item{c_t}{coalescent times.}

\item{beta}{growth rate.}

\item{t0}{time of most recent leave.}
}
\value{
rescaled waiting times.
}
\description{
Rescale a homogenous process to one with exponentially growing population.
}
