% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions_infer.R
\name{run_expansion_inference}
\alias{run_expansion_inference}
\title{Run rjmcmc inference on the provided phylogeny and supplied priors
This is the standard method for mcmc expansion inference.}
\usage{
run_expansion_inference(
  phy,
  priors,
  concentration = 1,
  thinning = 1,
  n_it = 1e+06,
  init = NULL
)
}
\arguments{
\item{phy}{phylogeny to run inference on}

\item{priors}{a list of prior likelihoods and sampling functions with names 'prior_i', 'prior_N', 'prior_t_mid_given_N', 'prior_t_mid_given_N.sample', 'prior_K_given_N', 'prior_K_given_N.sample', 'prior_t_given_N', 'prior_t_given_N.sample'. See `standard_priors` for details.}

\item{concentration}{concentration parameter for the dirichlet prior on expansion membership probabilities}

\item{thinning}{mcmc output thinning}

\item{n_it}{number of MCMC iterations}

\item{init}{initial state for the mcmc chain}
}
\value{
list with names 'phylo_preprocessed', the preprocessed phylogeny, and 'mcmc_out' a list of MCMC states
}
\description{
Run rjmcmc inference on the provided phylogeny and supplied priors
This is the standard method for mcmc expansion inference.
}
