% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run metropolis-hastings mcmc.}
\usage{
run_mcmc(
  model.log_lh,
  proposal.cond_log_lh,
  proposal.sampler,
  x0,
  n_it,
  verbose = FALSE
)
}
\arguments{
\item{x0}{initial value for the chain}

\item{n_it}{number of iterations}

\item{model.lh}{a function for the conditional model log-likelihood.}

\item{proposal.cond_lh}{a function for the conditional proposal log-likelihood \code{x_{t+1}} given \code{x_t}.}
}
\value{
A list of mcmc jumps
}
\description{
Run metropolis-hastings mcmc.
}
