% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sat_coalescent_loglh}
\alias{sat_coalescent_loglh}
\title{Compute likelihood of a realisation of an half/logistic growth coalescent.}
\usage{
sat_coalescent_loglh(sampling_times, coalescent_times, t0, r, N, t_max)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{coalescent_times}{times of coalescent events.}

\item{t0}{time of divergence.}

\item{N}{asymptotic population size.}

\item{growth}{rate.}
}
\value{
the log likelihood for the given parameters.
}
\description{
Compute likelihood of a realisation of an half/logistic growth coalescent.
}
