% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions_simulate.R
\name{simulate_expansion_phylo}
\alias{simulate_expansion_phylo}
\title{A phylogeny containing clonal expansions.
Note: The resulting phylogeny contains single child nodes denoting divergence events. use `collapse.singles` to remove these.}
\usage{
simulate_expansion_phylo(
  priors,
  sampling_times,
  concentration = 2,
  given = list(),
  collapse_singles = FALSE
)
}
\arguments{
\item{priors}{List of priors to simulate parameters from. See `standard priors` for details}

\item{sampling_times}{Vector sampling times, entries must be negative or 0}

\item{concentration}{Scalar concentration hyperparameter for dirichlet expansion model}

\item{given}{A list of variables with values given. Supported names: 'n_exp' - number of expansions, 'tip_colours' - tip expansion assignment, N' - Background population size, 'K' - Expansion carrying capacities, 't_mid' - Expansion times to midpoints, 'div_times' - Expansion divergence times, 'div_from' - parent lineages}

\item{collapse_singles}{Whether to remove (collapse) divergence event nodes. Default: FALSE}
}
\value{
list of: `tree` - The simulated genealogy, `params` - the simulated parameters for the process
}
\description{
A phylogeny containing clonal expansions.
Note: The resulting phylogeny contains single child nodes denoting divergence events. use `collapse.singles` to remove these.
}
