% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions_infer.R
\name{standard_priors}
\alias{standard_priors}
\title{Generates a list of priors of the form described in the paper to use with MCMC inference.}
\usage{
standard_priors(
  expansion_rate = 1,
  N_mean_log = 4,
  N_sd_log = 4,
  t_mid_rate = 10,
  K_sd_log = 1/2,
  exp_time_nu = 1/2,
  exp_time_kappa = 1/10
)
}
\arguments{
\item{expansion_rate}{corresponds to '\phi' in paper}

\item{N_mean_log}{corresponds to '\mu_{anc}' in paper}

\item{N_sd_log}{corresponds to '\sigma_{anc}' in paper}

\item{t_mid_rate}{corresponds to '\lambda_r' in paper}

\item{K_sd_log}{corresponds to '\sigma_{exp}' in paper}

\item{exp_time_nu}{corresponds to '\nu' in paper}

\item{exp_time_kappa}{corresponds to '\kappa' in paper}
}
\value{
a list of prior likelihoods and sampling functions with names 'prior_i', 'prior_i.sample', 'prior_N', 'prior_N.sample' ,prior_t_mid_given_N', 'prior_t_mid_given_N.sample', 'prior_K_given_N', 'prior_K_given_N.sample', 'prior_t_given_N', 'prior_t_given_N.sample'. See `standard_priors` for details.
}
\description{
Generates a list of priors of the form described in the paper to use with MCMC inference.
}
