% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_coal.R
\name{structured_coal.simulate}
\alias{structured_coal.simulate}
\title{Simulate expansion structured coalescent
Population size must be locally smooth around divergence time and must reach 0 at divergence time !!
Note: This simulation relies on inverse transform sampling and numerical inverses of Neg.rate.ints}
\usage{
structured_coal.simulate(
  sampling_times,
  colours,
  div_times,
  div_events,
  Neg.rates,
  Neg.rate.ints,
  div.from = NA
)
}
\arguments{
\item{sampling_times}{Nx1 times of leaves.}

\item{colours}{Nx1 colour assignment of leaves}

\item{div_times}{Kx1 divergence times for colours, neutral colour must diverge at -Inf. MUST BE IN DESCENDING ORDER}

\item{div_events}{Kx1 colour assignment for divergence times, i.e. which colour corresponds to which divergence time}

\item{Neg.rates}{Kx1 List of Functions (t)->(R+). 1/Neg(t) for each colour.}

\item{Neg.rate.ints}{Kx1 List of Functions (t, s)->(R+). int_{t}^{s} 1/Neg(tau),d tau for each colour}

\item{div.from}{Kx1 List of integers denoting parents of diverging lineages. If NA parents will be randomised}
}
\value{
A list consisting of the simulate coalescent times \code{coalescent_times} and the log-likelihood of the simulation \code{log_likelihood}.
}
\description{
Simulate expansion structured coalescent
Population size must be locally smooth around divergence time and must reach 0 at divergence time !!
Note: This simulation relies on inverse transform sampling and numerical inverses of Neg.rate.ints
}
