% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_utils.R
\name{transform_to_intervals}
\alias{transform_to_intervals}
\title{Precompute waiting times and lineage counts for intervals between events in a homogenous process.}
\usage{
transform_to_intervals(sampling_times, coalescent_times)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{coalescent_times}{times of coalescent events / internal nodes.}
}
\value{
a list of two entries \code{intervals} waiting times, \code{lineages} lineage counts.
}
\description{
Precompute waiting times and lineage counts for intervals between events in a homogenous process.
}
