% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_utils.R
\name{build_coal_tree}
\alias{build_coal_tree}
\title{Generate a Newick string representation of a tree corresponding to a realisation of a coalescent process.}
\usage{
build_coal_tree(
  sampling_times,
  coalescent_times,
  leaf_names = NULL,
  node_name_prefix = "N",
  terminate_string = TRUE
)
}
\arguments{
\item{sampling_times}{times of leaves.}

\item{coalescent_times}{times of coalescent events / internal nodes.}
}
\value{
a Newick string corresponding to the tree.
}
\description{
Generate a Newick string representation of a tree corresponding to a realisation of a coalescent process.
}
