% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coalescent_loglh}
\alias{coalescent_loglh}
\title{Compute likelihood of a realisation of a homogeneous coalescent.}
\usage{
coalescent_loglh(time_intervals, lineage_count, pop_size)
}
\arguments{
\item{time_intervals}{waiting times.}

\item{lineage_count}{number of lineages for each waiting time.}

\item{pop_size}{effective population size \code{Neg}.}
}
\value{
the log likelihood for the given parameters.
}
\description{
Compute likelihood of a realisation of a homogeneous coalescent.
}
