% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{plot_event_summary}
\alias{plot_event_summary}
\title{Converts MCMC output into two data frames, one conaining global model parameters and one containing expansion data}
\usage{
plot_event_summary(
  mcmc.df,
  event.df,
  which_br,
  pre,
  prior_N = NULL,
  prior_t_mid_given_N = NULL,
  prior_K_given_N = NULL,
  prior_t_given_N = NULL
)
}
\arguments{
\item{mcmc.df}{mcmc.df returned by mcmc2data.frame}

\item{event.df}{event.df returned by mcmc2data.frame}

\item{which_br}{Which branch to generate marginals for}

\item{pre}{Preprocessed phylogeny}

\item{prior_N}{(Optional) Background population size prior, mutually exclusive with passing list of priors. If either is supplied priors will be overlayed in plotting.}

\item{prior_t_mid_given_N}{(Optional) Time to midpoint  prior, mutually exclusive with passing list of priors. If either is supplied priors will be overlayed in plotting.}

\item{prior_K_given_N}{(Optional) Expansion time prior, mutually exclusive with passing list of priors. If either is supplied priors will be overlayed in plotting.}
}
\value{
a list of 3 plot panels
}
\description{
Converts MCMC output into two data frames, one conaining global model parameters and one containing expansion data
}
