% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{plot_tree_freq}
\alias{plot_tree_freq}
\title{Plots tree branch frequency histogram}
\usage{
plot_tree_freq(mcmc.df, event.df, pre, prior_tmid_given_N = NULL)
}
\arguments{
\item{mcmc.df}{mcmc.df returned by mcmc2data.frame}

\item{event.df}{event.df returned by mcmc2data.frame}

\item{pre}{preprocessed phylogeny}
}
\value{
a plot object
}
\description{
Plots tree branch frequency histogram
}
