% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run metropolis-hastings mcmc.}
\usage{
run_mcmc(model.lh, proposal.cond_lh, proposal.sampler, x0, n_it)
}
\arguments{
\item{model.lh}{a function for the conditional model likelihood.}

\item{proposal.cond_lh}{a function for the conditional proposal likelihood \code{x_{t+1}} given \code{x_t}.}

\item{x0}{initial value for the chain}

\item{n_it}{number of iterations}
}
\value{
A list of mcmc jumps
}
\description{
Run metropolis-hastings mcmc.
}
