% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_coal.R
\name{structured_coal.likelihood}
\alias{structured_coal.likelihood}
\title{Compute likelihood for preprocessed phylogeny}
\usage{
structured_coal.likelihood(
  phylo.preprocessed,
  div.MRCA.nodes,
  div_times,
  diverging.rates,
  diverging.sizes,
  neutral.size,
  times.extracted = NA,
  type = "Sat"
)
}
\arguments{
\item{phylo.preprocessed}{preprocessed phylogeny}

\item{div.MRCA.nodes}{labels of MRCA nodes of diverging lineages}

\item{div_times}{absolute divergence times}

\item{diverging.rates}{growth rates for diverging lineages}

\item{diverging.sizes}{asymptotic size for diverging lineages}

\item{neutral.size}{size of neutral phylogeny}

\item{times.extracted}{list returned by extract_lineage_times. Used instead of div.MRCA.nodes to avoid recomputing the tree partition.}
}
\value{
Log-likelihood
}
\description{
Compute likelihood for preprocessed phylogeny
}
