% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_coal.R
\name{structured_coal.preprocess_phylo}
\alias{structured_coal.preprocess_phylo}
\title{Preprocess phylogeny to speed up likelihood computation}
\usage{
structured_coal.preprocess_phylo(phy)
}
\arguments{
\item{phy}{phylogeny}
}
\value{
preprocessed phylogeny
}
\description{
Preprocess phylogeny to speed up likelihood computation
}
